+++ src/components/Playlist/hooks/useRenameSong.ts
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 const useRenameSong = () => {
   const [songObjEdited, setSongObjEdited] = useState<NoxMedia.Song>();
